+++ src/components/auth/SignUp.js
@@ -1,39 +1,12 @@
 import React, { useState } from "react";
-import {
-  Header,
-  Divider,
-  Form,
-  Container,
-  Loader,
-  Dimmer,
-} from "semantic-ui-react";
-import { useForm } from "react-hook-form";
-import axios from "axios";
+import { Header, Divider, Container, Loader, Dimmer } from "semantic-ui-react";
 import SidebarMarketing from "../marketing/Sidebar-Marketing.js";
 import Navigation from "../marketing/Navigation";
 import Footer from "../marketing/Footer";
+import SignUpForm from "./SignUpForm";
 
 const SignUp = (props) => {
-  const { register, handleSubmit, errors } = useForm();
   const [isLoading, setIsLoading] = useState(false);
-
-  const onSubmit = (data) => {
-    setIsLoading(true);
-    axios
-      .post(`${process.env.REACT_APP_BE_URL}/auth/signup`, data)
-      .then((res) => {
-        console.log("signup data", res);
-        // rather than pushing to the dashboard and setting the token here we should have a popup show that tells the user a confirmation email has been sent to them
-        // the confirmation route can then handle setting the token and pushing
-
-        // localStorage.setItem(`token`, res.data.payload);
-        // props.history.push(`/dashboard`);
-        setIsLoading(false);
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  };
 
   const googleAuth = () => {
     window.location = `${process.env.REACT_APP_BE_URL}/connect/google`;
@@ -56,51 +29,7 @@ const SignUp = (props) => {
               <Loader size="large">Loading</Loader>
             </Dimmer>
           ) : (
-            <Form onSubmit={handleSubmit(onSubmit)}>
-              <Form.Field>
-                <label>Username</label>
-                <input
-                  type="text"
-                  placeholder="Username"
-                  name="username"
-                  ref={register({ required: "Username is required." })}
-                />
-                {errors.username && <p>{errors.username.message}</p>}
-              </Form.Field>
-              <Form.Field>
-                <label>Email</label>
-                <input
-                  type="email"
-                  placeholder="Email"
-                  name="email"
-                  ref={register({ required: "Email is required." })}
-                />
-                {errors.email && <p>{errors.email.message}</p>}
-              </Form.Field>
-              <Form.Field>
-                <label>Password</label>
-                <input
-                  type="password"
-                  placeholder="Password"
-                  name="password"
-                  ref={register({
-                    required: "Password is required.",
-                    minLength: {
-                      value: 8,
-                      message: "Password must be at least 8 characters long.",
-                    },
-                  })}
-                />
-                {errors.password && <p>{errors.password.message}</p>}
-              </Form.Field>
-              {/*<Form.Field>
-              <label>Repeat Password</label>
-              <input type="password" placeholder="Repeat Password" />
-            </Form.Field>*/}
-              <button type="submit" className="ui button blue">
-                Submit
-              </button>
-            </Form>
+            <SignUpForm setIsLoading={setIsLoading} />
           )}
         </div>
         <p>
